:recycle: [Refactor] friend 도메인 분리
