+++ src/main/java/com/gamegoo/controller/member/FriendController.java
@@ -2,8 +2,8 @@ package com.gamegoo.controller.member;
 
 import com.gamegoo.apiPayload.ApiResponse;
 import com.gamegoo.converter.MemberConverter;
-import com.gamegoo.domain.Friend;
-import com.gamegoo.domain.FriendRequests;
+import com.gamegoo.domain.friend.Friend;
+import com.gamegoo.domain.friend.FriendRequests;
 import com.gamegoo.dto.member.MemberResponse;
 import com.gamegoo.service.member.FriendService;
 import com.gamegoo.util.JWTUtil;
