+++ src/main/java/com/gamegoo/repository/friend/FriendRepository.java
@@ -1,6 +1,6 @@
 package com.gamegoo.repository.friend;
 
-import com.gamegoo.domain.Friend;
+import com.gamegoo.domain.friend.Friend;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
