+++ src/main/java/com/gamegoo/repository/friend/FriendRequestsRepository.java
@@ -1,6 +1,6 @@
 package com.gamegoo.repository.friend;
 
-import com.gamegoo.domain.FriendRequests;
+import com.gamegoo.domain.friend.FriendRequests;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FriendRequestsRepository extends JpaRepository<FriendRequests, Long> {
