+++ src/main/java/com/gamegoo/util/MemberUtils.java
